+++ src/jsMain/kotlin/dev/inmo/jsuikit/elements/Dropdown.kt
@@ -3,10 +3,26 @@ package dev.inmo.jsuikit.elements
 import androidx.compose.runtime.Composable
 import dev.inmo.jsuikit.buildAndAddAttribute
 import dev.inmo.jsuikit.modifiers.*
+import dev.inmo.jsuikit.types.DropdownOptions
 import dev.inmo.jsuikit.utils.Milliseconds
 import org.jetbrains.compose.web.dom.*
 import org.w3c.dom.HTMLDivElement
 
+@Composable
+fun Dropdown(
+    vararg modifiers: UIKitModifier,
+    dropdownOptions: DropdownOptions,
+    attributesCustomizer: AttrBuilderContext<HTMLDivElement> = {},
+    contentBuilder: ContentBuilder<HTMLDivElement>
+) {
+    Div(
+        {
+            include(UIKitDropdown(dropdownOptions), *modifiers)
+            attributesCustomizer()
+        },
+        contentBuilder
+    )
+}
 @Composable
 fun Dropdown(
     vararg modifiers: UIKitModifier,
@@ -26,21 +42,24 @@ fun Dropdown(
 ) {
     Div(
         {
-            buildAndAddAttribute("uk-dropdown") {
-                "toggle" to toggle
-                "pos" to pos
-                "mode" to mode
-                "delayShow" to delayShow
-                "delayHide" to delayHide
-                "boundary" to boundary
-                "boundaryAlign" to boundaryAlign
-                "flip" to flip
-                "offset" to offset
-                "animation" to animation
-                "duration" to duration
-            }
-            classes("uk-dropdown")
-            include(*modifiers)
+            include(
+                UIKitDropdown(
+                    DropdownOptions(
+                        toggle = toggle,
+                        pos = pos,
+                        mode = mode,
+                        delayShow = delayShow,
+                        delayHide = delayHide,
+                        boundary = boundary,
+                        boundaryAlign = boundaryAlign,
+                        flip = flip,
+                        offset = offset,
+                        animation = animation,
+                        duration = duration
+                    )
+                ),
+                *modifiers
+            )
             attributesCustomizer()
         },
         contentBuilder