+++ src/jsMain/kotlin/dev/inmo/jsuikit/utils/AttributeBuilder.kt
@@ -2,26 +2,24 @@ package dev.inmo.jsuikit.utils
 
 class AttributeBuilder (
     val attributeName: String,
-    val skipNullValues: Boolean = true,
-    private val parametersPreset: MutableMap<String, String?> = mutableMapOf()
+    val parametersBuilder: ParametersBuilder
 ) {
-
-    fun add(k: String, v: Any? = null) {
-        if (v != null || !skipNullValues) {
-            parametersPreset[k] = v ?.toString()
-        }
+    fun build(): Pair<String, String> = parametersBuilder.build().run {
+        Pair(
+            attributeName, toList().joinToString(";") {
+                "${it.first}${it.second ?.let { ": $it" } ?: ""}"
+            }
+        )
     }
-    infix fun String.to(value: Any?) = add(this, value)
-    operator fun String.unaryPlus() = add(this, null)
-
-    fun build(): Pair<String, String> = Pair(
-        attributeName, parametersPreset.toList().joinToString(";") {
-            "${it.first}${it.second ?.let { ": $it" } ?: ""}"
-        }
-    )
 }
 
-inline fun buildAttribute(attributeName: String, skipNullValues: Boolean = true, block: AttributeBuilder.() -> Unit) = AttributeBuilder(
+inline fun buildAttribute(
+    attributeName: String,
+    skipNullValues: Boolean = true,
+    block: ParametersBuilder.() -> Unit
+) = AttributeBuilder(
     attributeName,
-    skipNullValues
-).apply(block).build()
+    ParametersBuilder(skipNullValues)
+).apply {
+    parametersBuilder.block()
+}.build()